backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.7
pandas>=2.2
numpy>=1.26
python-dotenv>=1.0
openai>=1.30
rapidfuzz>=3.9
beautifulsoup4>=4.12
lxml>=5.2
pdfminer.six>=20231228
requests>=2.32
tenacity>=8.2
